Make tool installation handler modular
